Change service to use timeout instead of last-request tracking.
This means that the service will be waiting for requests in the queue
for a specified amount of time, after this time passes and the queue
remains empty the service will start to shutdown. After the shutdown
has been completed any new requests will end up spinning a new server.
The service will re-request requests that arrived between starting to
shutdown and its completion.