Fix crash on recycler view update race
RV was crashing when clear queue was clicked when the UI was janky.
This was happening because we were executing the LiveData
transformations off the main-thread and then submitting the results
back on the main-thread without preserving initial update arrival
order, which in turns broke our assumptions about the way the data
changed (in HistoryAdapter.notifyHistoryChanged()).